requirements.txt
aiogram>=3.4.1,<4.0.0
aiosqlite>=0.19.0,<1.0.0
aiohttp>=3.9.1,<4.0.0
python-dotenv>=1.0.0,<2.0.0

# TON
pytoniq>=0.11.4,<1.0.0
pytoniq-core>=0.1.6,<1.0.0
tonsdk>=0.0.9,<1.0.0

# Crypto & Wallet
bip-utils>=2.9.3,<3.0.0
cryptography>=41.0.7,<45.0.0

# TRON
tronpy>=0.5.0,<1.0.0
